Move cloud spec validation into separate function to follow existing code structure
